cpp::09::ex02 Done

+++ cpp09/ex02/PmergeMe.hpp
@@ -6,7 +6,7 @@
 /*   By: ouakrad <ouakrad@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/05 15:36:20 by ouakrad           #+#    #+#             */
-/*   Updated: 2024/02/06 17:41:49 by ouakrad          ###   ########.fr       */
+/*   Updated: 2024/02/07 15:35:37 by ouakrad          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 #include <deque>
 #include <sstream>
 
-void displaySequence(const std::vector<int> &sequence, const std::string &message);
+void displaySequence(const std::vector<long> &sequence, const std::string &message);
 void displayTiming(const std::string &containerName, const clock_t &startTime, const clock_t &endTime);
 void fordJohnsonMergeInsertionSort(std::vector<int> &c, int start, int end);
 void performSorting(std::vector<int> sequence,std::deque<int> sequenceDeque);
